Revert "Dev"

+++ app/src/main/java/hr/foi/air1719/personaltracker/Helper.java
@@ -4,6 +4,9 @@ import android.content.Context;
 import android.location.Location;
 import android.net.ConnectivityManager;
 import android.text.TextUtils;
+
+import java.net.InetAddress;
+import java.net.UnknownHostException;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.Date;
@@ -48,20 +51,20 @@ public class Helper {
         return false;
     }
 
-    public static float CalculateDistance(Location startPoint, Location endPoint)
+    public static double CalculateDistance(Location startPoint, Location endPoint)
     {
-        return (startPoint.distanceTo(endPoint) / ((float)1000));
+        return (startPoint.distanceTo(endPoint)/1000);
     }
 
-    public static float CalculateAvgSpeed(Date startDate, Date endDate, double KM)
+    public static double CalculateAvgSpeed(Date startDate, Date endDate, double KM)
     {
         try {
             long different = endDate.getTime() - startDate.getTime();
 
-            float diffHours = (float)((float)different / (float)(60 * 60 * 1000));
+            double diffHours = (different / (60 * 60 * 1000));
             if(diffHours==0) return 0;
 
-            return (float)(KM / diffHours);
+            return (KM / diffHours);
         }
         catch (Exception E)
         {
@@ -69,4 +72,5 @@ public class Helper {
         }
         return -1;
     }
+
 }
